requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.3
django-filter==22.1
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
multidict==6.0.2
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
zipp==3.10.0
